+++ client/app/src/main/java/hu/bme/aut/rentapp/models/ContractModelPost.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 import java.util.*
 
 data class ContractModelPost(
-    @SerializedName("id") val id : Int? = null,
-    @SerializedName("vehicles") val vehicles : List<VehicleModel>? = null,
+    @SerializedName("id") val id: Long? = null,
+    @SerializedName("vehicles") val vehicles: List<VehicleModel>? = null,
     @SerializedName("startDate") val startDate: Date? = null,
     @SerializedName("endDate") val endDate: Date? = null,
 )
