+++ client/app/src/main/java/hu/bme/aut/rentapp/models/VehicleModel.kt
@@ -1,10 +1,10 @@
 package hu.bme.aut.rentapp.models
 
 data class VehicleModel(
-    val id : Int? = null,
+    val id : Long? = null,
     val name : String? = null,
     val brand : String? = null,
-    val price : Double? = null,
+    val price : Int? = null,
     val description : String? = null,
     val category: String? = null,
 )
